+++ src/data/protocols/http/http-post-client.ts
@@ -1,5 +1,5 @@
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
-import { HttpResponse } from './http-response'
+import { HttpResponse } from '.'
 
 export type HttpPostParams<T> = {
   url: string
