+++ src/data/protocols/http/index.ts
@@ -1,0 +1,2 @@
+export * from './http-post-client'
+export * from './http-response'
