+++ src/data/test/mock-http-client.ts
@@ -1,7 +1,5 @@
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
-import { HttpPostClient, HttpPostParams } from '@/data/protocols/http/http-post-client'
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-import { HttpResponse, HttpStatusCode } from '@/data/protocols/http/http-response'
+import { HttpPostClient, HttpPostParams, HttpResponse, HttpStatusCode } from '@/data/protocols/http'
 
 export class HttpPostClientSpy<T, R> implements HttpPostClient<T, R> {
   url?: string
