+++ src/domain/errors/index.ts
@@ -1,0 +1,2 @@
+export * from './invalid-credentials-error'
+export * from './unexpected-error'
